+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/BlueJewelOnly.java
@@ -1,5 +1,8 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode;
 
+/**
+ * Created by anikaitsingh on 12/9/17.
+ */
 
 import android.graphics.Color;
 
@@ -9,10 +12,8 @@ import com.qualcomm.robotcore.hardware.ColorSensor;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.Utilities.Robot;
-
-@Autonomous(name = "Red: Jewel Only", group = "Sensor")
-public class RedJewelOnly extends LinearOpMode {
+@Autonomous(name = "Blue: Jewel Only", group = "Sensor")
+public class BlueJewelOnly extends LinearOpMode {
 
     ColorSensor sensorColor;
     Robot robot;
@@ -63,9 +64,9 @@ public class RedJewelOnly extends LinearOpMode {
 
         double power;
         if (red) {
-            power = .2;
-        } else {
             power = -.2;
+        } else {
+            power = .2;
         }
 
         //for testing purposes
