+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/DriveTime.java
@@ -1,11 +1,11 @@
-package org.firstinspires.ftc.teamcode.Tests;
+package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.Utilities.Robot;
+import org.firstinspires.ftc.teamcode.Robot;
 
 @Autonomous(name = "Driving Time", group = "Test")
 public class DriveTime extends LinearOpMode{
