+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Logging.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Utilities;
+package org.firstinspires.ftc.teamcode;
 
 import android.util.Log;
