+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MoveOneRotation.java
@@ -1,11 +1,11 @@
-package org.firstinspires.ftc.teamcode.Tests;
+package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
-import org.firstinspires.ftc.teamcode.Utilities.Logging;
-import org.firstinspires.ftc.teamcode.Utilities.Robot;
+import org.firstinspires.ftc.teamcode.Logging;
+import org.firstinspires.ftc.teamcode.Robot;
 
 /**
  * Created by anikaitsingh on 11/26/17.
