+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ParkAuton.java
@@ -1,10 +1,8 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
-
-import org.firstinspires.ftc.teamcode.Utilities.Robot;
 
 /**
  * Created by anikaitsingh on 11/1/17.
