+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SensorColor.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Utilities;
+package org.firstinspires.ftc.teamcode;
 
 import android.graphics.Color;
 import android.util.Log;
