+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Teleop.java
@@ -1,10 +1,8 @@
-package org.firstinspires.ftc.teamcode.Utilities;
+package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
-
-import org.firstinspires.ftc.teamcode.Utilities.Robot;
 
 @TeleOp(name = "tele")
 public class Teleop extends OpMode {
