+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TickCounter.java
@@ -1,10 +1,10 @@
-package org.firstinspires.ftc.teamcode.Tests;
+package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
-import org.firstinspires.ftc.teamcode.Utilities.Robot;
+import org.firstinspires.ftc.teamcode.Robot;
 
 @Autonomous(name = "Tick Counter test", group = "Test")
